requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
google-cloud-storage
sqlalchemy
psycopg[binary]
alembic
google-cloud-secret-manager
